add prefix() to folders (respecting YOTTA_PREFIX env var), move folders into lib; use prefix when linking
